Compres to one
UserViewAllUser.java

users by type and users status
with drop downl
